Fix Trace Consumer Bug on Busy Load/Store When a Slot is Full
Traces could be written by device-side trace kernels, even after a host-side trace consumer decided to consume data ('commit > SLOTS_SIZE - 32'), if 'alloc > commit' at the time of the decision. In this case, trace kernels tries to write traces to the slot, which are not read by the consumer because max read 'count == commit'.